move legacy TextWithCopy
